% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfunction.R
\name{blackwhite_pal}
\alias{blackwhite_pal}
\title{costum color package base function}
\usage{
blackwhite_pal(palette = "blackwhite", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{choose a pellet from the 'colorlist'}

\item{alpha}{sets the transparency}

\item{reverse}{if set to true, the order of the colorlist is reversed}
}
\description{
costum color package base function
}
\examples{
show_col(blackwhite_pal("blackwhite")(7))
filled.contour(volcano,color.palette = blackwhite_pal("blackwhite"), asp=1)

}
