% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfunction.R
\name{color_pallete_blackwhite}
\alias{color_pallete_blackwhite}
\title{color palette for ggplot2}
\usage{
color_pallete_blackwhite(
  ...,
  palette = "blackwhite",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{...}{allows for additional arguments to scale_color_gradient, is not a requirement to fill in}

\item{palette}{choose a pellet from the 'colorlist'}

\item{discrete}{allows to choose a discrete color palette or not}

\item{alpha}{sets the transparency}

\item{reverse}{if set to true, the order of the colorlist is reversed}
}
\description{
color palette for ggplot2
}
\examples{
ggplot(mtcars, aes(hp, mpg)) +
geom_point(aes(color = factor(carb)), size = 4) +
 color_pallete_blackwhite()


}
