% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{correlation_function}
\alias{correlation_function}
\title{quick correlation function allowing to quickly analyse if the x and y variable are correlated}
\usage{
correlation_function(data1 = quick_select_data)
}
\arguments{
\item{data1}{input data, usually the "quick_select_data" from the "quick_select_function"}
}
\description{
quick correlation function allowing to quickly analyse if the x and y variable are correlated
}
\examples{
correlation_function()
correlation_function(data1 = quick_select_data)

}
