% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorfunction.R
\name{fill_pallete_blackwhite}
\alias{fill_pallete_blackwhite}
\title{fill palette}
\usage{
fill_pallete_blackwhite(
  ...,
  palette = "blackwhite",
  discrete = TRUE,
  alpha = 1,
  reverse = TRUE
)
}
\arguments{
\item{...}{allows for additional arguments to scale_color_gradient, is not a requirement to fill in}

\item{palette}{choose a pellet from the 'colorlist'}

\item{discrete}{allows to choose a discrete color palette or not}

\item{alpha}{sets the transparency}

\item{reverse}{if set to true, the order of the colorlist is reversed}
}
\description{
fill palette
}
\examples{
ggplot(mtcars) + geom_bar(aes(x = hp, fill = mpg)) + fill_pallete_blackwhite()

}
