% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_functions.R
\name{quick_Select_function}
\alias{quick_Select_function}
\title{quick select function to form data frame for other quick functions}
\usage{
quick_Select_function(
  groupselect = data$gender,
  xselect = data$age,
  yselect = data$hight
)
}
\arguments{
\item{groupselect}{selects a group from a data frame}

\item{xselect}{selects values for x axis in quick graph functions}

\item{yselect}{selects values for y axis in quick graph function}
}
\description{
quick select function to form data frame for other quick functions
}
\examples{
quick_select_function(data$gender, data$age, data$hight)

}
